requirements.txt
azure-core==1.29.7
azure-storage-blob==12.19.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.1
idna==3.6
isodate==0.6.1
numpy==1.26.3
pandas==2.2.0
pycparser==2.21
pyspnego==0.10.2
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
smbprotocol==1.12.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
